+++ etna/transforms/__init__.py
@@ -27,6 +27,8 @@ from etna.transforms.feature_selection import GaleShapleyFeatureSelectionTransfo
 from etna.transforms.feature_selection import MRMRFeatureSelectionTransform
 from etna.transforms.feature_selection import TreeFeatureSelectionTransform
 from etna.transforms.math import AddConstTransform
+from etna.transforms.math import BinaryOperationTransform
+from etna.transforms.math import BinaryOperator
 from etna.transforms.math import BoxCoxTransform
 from etna.transforms.math import DifferencingTransform
 from etna.transforms.math import ExogShiftTransform