+++ tests/test_transforms/test_inference/test_inverse_transform.py
@@ -9,6 +9,7 @@ from sklearn.tree import DecisionTreeRegressor
 from etna.analysis import StatisticsRelevanceTable
 from etna.models import ProphetModel
 from etna.transforms import AddConstTransform
+from etna.transforms import BinaryOperationTransform
 from etna.transforms import BoxCoxTransform
 from etna.transforms import ChangePointsLevelTransform
 from etna.transforms import ChangePointsSegmentationTransform
@@ -147,6 +148,18 @@ class TestInverseTransformTrainSubsetSegments:
             # math
             (AddConstTransform(in_column="target", value=1, inplace=False), "regular_ts"),
             (AddConstTransform(in_column="target", value=1, inplace=True), "regular_ts"),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="positive"
+                ),
+                "ts_with_exog",
+            ),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="new_col"
+                ),
+                "ts_with_exog",
+            ),
             (LagTransform(in_column="target", lags=[1, 2, 3]), "regular_ts"),
             (
                 LambdaTransform(in_column="target", transform_func=lambda x: x + 1, inplace=False),
@@ -345,6 +358,18 @@ class TestInverseTransformFutureSubsetSegments:
             (AddConstTransform(in_column="target", value=1, inplace=False), "regular_ts"),
             (AddConstTransform(in_column="target", value=1, inplace=True), "regular_ts"),
             (AddConstTransform(in_column="positive", value=1, inplace=True), "ts_with_exog"),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="positive"
+                ),
+                "ts_with_exog",
+            ),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="new_col"
+                ),
+                "ts_with_exog",
+            ),
             (LagTransform(in_column="target", lags=[1, 2, 3]), "regular_ts"),
             (
                 LambdaTransform(in_column="target", transform_func=lambda x: x + 1, inplace=False),
@@ -578,6 +603,20 @@ class TestInverseTransformTrainNewSegments:
                 {},
             ),
             (AddConstTransform(in_column="target", value=1, inplace=True), "regular_ts", {"change": {"target"}}),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="positive"
+                ),
+                "ts_with_exog",
+                {"change": {"positive"}},
+            ),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="new_col"
+                ),
+                "ts_with_exog",
+                {},
+            ),
             (
                 LagTransform(in_column="target", lags=[1, 2, 3], out_column="res"),
                 "regular_ts",
@@ -876,6 +915,20 @@ class TestInverseTransformFutureNewSegments:
             ),
             (AddConstTransform(in_column="target", value=1, inplace=True), "regular_ts", {}),
             (AddConstTransform(in_column="positive", value=1, inplace=True), "ts_with_exog", {"change": {"positive"}}),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="weekday", operator="+", out_column="positive"
+                ),
+                "ts_with_exog",
+                {"change": {"positive"}},
+            ),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="weekday", operator="+", out_column="new_col"
+                ),
+                "ts_with_exog",
+                {},
+            ),
             (
                 LagTransform(in_column="target", lags=[1, 2, 3], out_column="res"),
                 "regular_ts",
@@ -1338,6 +1391,20 @@ class TestInverseTransformFutureWithTarget:
             ),
             (AddConstTransform(in_column="target", value=1, inplace=True), "regular_ts", {"change": {"target"}}),
             (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="positive"
+                ),
+                "ts_with_exog",
+                {"change": {"positive"}},
+            ),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="target", operator="+", out_column="new_col"
+                ),
+                "ts_with_exog",
+                {},
+            ),
+            (
                 LagTransform(in_column="target", lags=[1, 2, 3], out_column="res"),
                 "regular_ts",
                 {},
@@ -1718,6 +1785,20 @@ class TestInverseTransformFutureWithoutTarget:
             ),
             (AddConstTransform(in_column="target", value=1, inplace=True), "regular_ts", {}),
             (AddConstTransform(in_column="positive", value=1, inplace=True), "ts_with_exog", {"change": {"positive"}}),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="weekday", operator="+", out_column="positive"
+                ),
+                "ts_with_exog",
+                {"change": {"positive"}},
+            ),
+            (
+                BinaryOperationTransform(
+                    left_column="positive", right_column="weekday", operator="+", out_column="new_col"
+                ),
+                "ts_with_exog",
+                {},
+            ),
             (
                 LagTransform(in_column="target", lags=[1, 2, 3], out_column="res"),
                 "regular_ts",